Fix -Wtautological-constant-out-of-range-compare warnings: result of comparison of constant xxx with expression of type 'yyyy' is always false

+++ Code/Scripts/Mission02.cpp
@@ -3452,7 +3452,7 @@ DECLARE_SCRIPT (M02_Nod_Soldier, "Area_Number:int,Area_Officer:int,Pre_Placed:in
 	{
 		// If an objective shouts, respond to it, letting it know the area is not cleared.
 
-		if (sound.Type == 1000)
+		if (static_cast<int>(sound.Type) == 1000)
 		{
 			if (sound.Creator)
 			{
@@ -3462,7 +3462,7 @@ DECLARE_SCRIPT (M02_Nod_Soldier, "Area_Number:int,Area_Officer:int,Pre_Placed:in
 
 		// If a zone shouts to return to home points, do so.
 
-		else if ((sound.Type > 1049) && (sound.Type < 1100) && (!no_return_home))
+		else if ((static_cast<int>(sound.Type) > 1049) && (static_cast<int>(sound.Type) < 1100) && (!no_return_home))
 		{
 			int area_id = (sound.Type - 1050);
 			int area_num = Get_Int_Parameter("Area_Number");
@@ -4335,12 +4335,12 @@ DECLARE_SCRIPT (M02_GDI_Soldier, "Area_ID:int, Soldier_Type=0:int")
 
 	void Sound_Heard (GameObject * obj, const CombatSound & sound) override
 	{
-		if ((sound.Type > 1099) && (sound.Type < 1199))
+		if ((static_cast<int>(sound.Type) > 1099) && (static_cast<int>(sound.Type) < 1199))
 		{
 			// Make sure this is the sound this unit is supposed to react to.
 
 			int area_id = Get_Int_Parameter("Area_ID");
-			if ((sound.Type - 1100) == area_id)
+			if ((static_cast<int>(sound.Type) - 1100) == area_id)
 			{
 				stop_following = true;
 				switch (area_id)
@@ -4498,7 +4498,7 @@ DECLARE_SCRIPT (M02_Stationary_Vehicle,"Area_ID:int")
 	{
 		// If an objective shouts, respond to it, letting it know the area is not cleared.
 
-		if (sound.Type == 1000)
+		if (static_cast<int>(sound.Type) == 1000)
 		{
 			if (sound.Creator)
 			{
@@ -4694,7 +4694,7 @@ DECLARE_SCRIPT (M02_Nod_Apache, "Area_ID:int")
 
 	void Sound_Heard (GameObject * obj, const CombatSound & sound) override
 	{
-		if (sound.Type == 1000)
+		if (static_cast<int>(sound.Type) == 1000)
 		{
 			if (sound.Creator)
 			{
@@ -4759,7 +4759,7 @@ DECLARE_SCRIPT (M02_Nod_Vehicle, "Area_ID:int")
 
 	void Sound_Heard (GameObject * obj, const CombatSound & sound) override
 	{
-		if (sound.Type == 1000)
+		if (static_cast<int>(sound.Type) == 1000)
 		{
 			if (sound.Creator)
 			{
@@ -4950,7 +4950,7 @@ DECLARE_SCRIPT (M02_Nod_Sakura, "Area_ID:int")
 
 	void Sound_Heard (GameObject * obj, const CombatSound & sound) override
 	{
-		if (sound.Type == 1000)
+		if (static_cast<int>(sound.Type) == 1000)
 		{
 			if (sound.Creator)
 			{
